improved demo stats service logic, updated kotlin/ktor versions & api changes
