+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
-    ext.kotlin_version = "$kotlin_version"
-    ext.ktor_version = "$ktor_version"
+    ext.kotlin_version = "1.5.30"
+    ext.ktor_version = "1.6.3"
     repositories {
         mavenLocal()
         mavenCentral()
@@ -38,9 +38,9 @@ repositories {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    implementation "org.jetbrains.exposed:exposed:0.12.1"
-    implementation "com.zaxxer:HikariCP:2.7.8"
-    implementation "org.postgresql:postgresql:42.2.2"
+    implementation 'org.jetbrains.exposed:exposed:0.17.13'
+    implementation 'com.zaxxer:HikariCP:5.0.0'
+    implementation 'org.postgresql:postgresql:42.2.23.jre7'
     implementation "io.ktor:ktor-server-netty:$ktor_version"
     implementation "ch.qos.logback:logback-classic:$logback_version"
     implementation "io.ktor:ktor-server-core:$ktor_version"
@@ -51,7 +51,7 @@ dependencies {
     implementation 'com.dropbox.core:dropbox-core-sdk:3.1.3'
     implementation "io.ktor:ktor-client-gson:$ktor_version"
     implementation 'net.dv8tion:JDA:4.1.1_144'
-    implementation "com.squareup.okhttp3:okhttp:4.6.0"
+    implementation 'com.squareup.okhttp3:okhttp:4.9.1'
     testImplementation "io.ktor:ktor-server-tests:$ktor_version"
     implementation "ch.qos.logback:logback-classic:$logback_version"
     implementation "io.ktor:ktor-client-logging:$ktor_version"
