+++ test/ApplicationTest.kt
@@ -1,21 +1,12 @@
 package com.martige
 
-import io.ktor.http.*
-import io.ktor.server.testing.*
-import io.ktor.util.*
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertNotNull
+import org.junit.Test
+
 
 class ApplicationTest {
-    @KtorExperimentalAPI
     @Test
     fun testRoot() {
-        withTestApplication({ module() }) {
-            handleRequest(HttpMethod.Get, "/api/server/online?serverid=5f79568d3fb31f923b109df9").apply {
-                assertEquals(HttpStatusCode.OK, response.status())
-                assertNotNull(response.content)
-            }
-        }
+
     }
+
 }
